Pagination Project- React Redux Toolkit
The MovieDetails app is a project where users can search for movies and view their details. The app allows users to input a movie name and retrieves information such as the title, release date, plot, cast, and poster. It fetches this data from an external API like OMDB (Open Movie Database). The main features of the app include:

    Search Functionality: Users can search for movies by title.
    Movie Details: Displays detailed information about the selected movie, including plot, actors, and ratings.
    UI Elements: Clean and user-friendly design for displaying movie information, with separate components for the search bar and movie details.
    API Integration: Fetches data from an external movie database API.
    Error Handling: Provides error messages if a movie is not found or there is an issue with the API.

It is a straightforward app with the primary goal of showcasing movie data based on user input.

Hosted: https://movie-details-phi.vercel.app/